+++ src/main/java/link/locutus/discord/util/PW.java
@@ -1111,7 +1111,7 @@ public final class PW {
         while (iter.hasNext()) {
             Map.Entry<ResourceType, Double> entry = iter.next();
             if (entry.getValue() < 0) {
-                negativeTotal += Locutus.imp().getTradeManager().getHigh(entry.getKey()) * entry.getValue().doubleValue() * -1;
+                negativeTotal += Locutus.imp().getTradeManager().getHigh(entry.getKey()) * entry.getValue() * -1;
                 iter.remove();
             }
         }
